requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
clickhouse-connect==0.7.0
pydantic==2.6.0
pydantic-settings==2.1.0
slowapi==0.1.9
pyyaml==6.0.1
requests==2.31.0